refactor: move `module_env` to the root of the crate, remove `environ` module
