--- frontend-wasm/src/environ/mod.rs
@@ -1,6 +1,0 @@
-//! Wasm to Miden IR translation environment
-
-// TODO: "inline" module_env::* here
-mod module_env;
-
-pub use crate::environ::module_env::{ModuleEnvironment, ModuleInfo};
