+++ frontend-wasm/src/module_translator.rs
@@ -1,8 +1,8 @@
 //! Translation skeleton that traverses the whole WebAssembly module and call helper functions
 //! to deal with each part of it.
 
-use crate::environ::ModuleEnvironment;
 use crate::error::WasmResult;
+use crate::module_env::ModuleEnvironment;
 use crate::sections_translator::{
     parse_data_section, parse_element_section, parse_function_section, parse_global_section,
     parse_import_section, parse_memory_section, parse_name_section, parse_type_section,
